requirements.txt
Telethon
pytz
python-dotenv